Close #12
There was an issue with Iris where if you 3D/Haptic Touched on a message to bring up the reactions, you would no longer be able to return to the conversation list by pressing the back button. I spent hours tracking this down to undefined behaviour in a ton of private UIKit methods, mostly related to `UINavigationController` and transitions.
Fast-forward 11 hours, I noticed that this was caused by the code which adds the Iris button and dimming view to the navigation item. I noticed that the issue would be fixed if I removed them, and it would remain fixed if I added something else in their place. So, I then checked the code to see if there was anything relating to layout/animations/transitions, but all I could see was constraints code. I commented this code out, and it was fixed… hold on, what? I investigated further and found that the issue was that I was constraining to `window`, rather than to `window?.rootViewController?.view`.
So… let me get this straight: Constraining to `window` in a navigation bar item causes undefined behaviour with `UINavigationController` transitions? I'm guessing this is a bug with UIKit… or have broken some cardinal rule I'm unaware of? Who knows…